requirements.txt
discord.py
random
os
yuotube-dl
pyowm
aiohttp
websockets
chardet